requirements.txt
amqp==2.6.1
asgiref==3.7.2
Babel==2.13.0
billiard==3.6.4.0
cairocffi==1.6.1
CairoSVG==2.7.1
celery==4.4.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.22
django-localflavor==4.0
django-rosetta==0.9.9
flower==0.9.3
html5lib==1.1
idna==3.4
importlib-metadata==4.13.0
kombu==4.6.11
Pillow==9.5.0
polib==1.2.0
pycparser==2.21
pyphen==0.14.0
python-stdnum==1.19
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tornado==5.1.1
typing-extensions==4.7.1
urllib3==2.0.7
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
zipp==3.15.0
